replicate sources
