+++ maven/apps/kotlin-portlet/src/main/kotlin/com/liferay/blade/samples/portlet/kotlin/KotlinGreeterActionCommandKt.kt
@@ -1,18 +1,20 @@
 /**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
-
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
-
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
+ * Copyright 2000-present Liferay, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-package com.liferay.blade.samples.portlet.kotlinactioncommand
+package com.liferay.blade.samples.portlet.kotlin
 
 import com.liferay.portal.kernel.log.LogFactoryUtil
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand
@@ -32,7 +34,7 @@ import org.osgi.service.component.annotations.Component
 @Component(
 	immediate = true,
 	property = arrayOf(
-		"javax.portlet.name=com_liferay_blade_samples_portlet_kotlinactioncommand_KotlinGreeterPortletKt",
+		"javax.portlet.name=KotlinGreeterPortletKt",
 		"mvc.command.name=greet"
 	),
 	service = arrayOf(MVCActionCommand::class))
