--- maven/apps/kotlin-portlet/src/main/kotlin/com/liferay/blade/samples/portlet/kotlinactioncommand/KotlinGreeterPortletKt.kt
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
-
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
-
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.blade.samples.portlet.kotlinactioncommand
-
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet
-
-import javax.portlet.Portlet
-
-import org.osgi.service.component.annotations.Component
-
-/**
- * @author Liferay
- */
-@Component(
-	immediate = true,
-	property = arrayOf(
-		"com.liferay.portlet.css-class-wrapper=portlet-greeter",
-		"com.liferay.portlet.display-category=category.sample",
-		"com.liferay.portlet.instanceable=true",
-		"javax.portlet.display-name=Kotlin Greeter Portlet",
-		"javax.portlet.init-param.template-path=/",
-		"javax.portlet.init-param.view-template=/view.jsp",
-		"javax.portlet.security-role-ref=power-user,user"
-	),
-	service = arrayOf(Portlet::class))
-class KotlinGreeterPortletKt : MVCPortlet()
