+++ maven/apps/kotlin-portlet/src/main/resources/content/Language.properties
@@ -1,1 +1,1 @@
-blade_portlet_kotlinactioncommand_KotlinGreeterPortlet.caption=Hello from BLADE!
+blade_portlet_kotlinactioncommand_KotlinGreeterPortlet.caption=Hello from Kotlin Portlet!
